This is a working build that makes all of the necessary parts to deploy to the Maven Repository Central
I still have this setup so we have to manually sign all of the jar
files in the target directory via gpg to make the ascii armor signature
files  *.asc

Also, we have to manually copy the pom.xml into the target directory as
itext-4.2.0.pm 

and then run

$ gpg -ab itext-4.2.0.pom 

to create the itext-4.2.0.asc file 

I have to get the maven-gpg-plugin to work properly.

after signing manually, we can run mvn repository:bundle-create
to wrap up all of the bundle which can be uploaded to sonatype.
I want to also work on making the release process work.