Feature: Include tag names if available, allow alternate text for an element
Also, add <br /> tags where line breaks occurred.

We recognize the replacement text that can be entered directly into a PDF
Artifact, and render it instead of the content of that element if it is present.
unfortunately the replacement text does not have a meaningful position on the screen
so read speaker highlighting will probably be pretty odd.