Added OSGi mappings for all projects
This will add `Export-Package` and `Import-Package` entries to the
generated JAR files' MANIFEST.MF files, allowing to deploy the libraries
as OSGi bundles on application servers supporting OSGi.

The Bouncycastle dependencies have been marked as optional, since they
are only required for some use cases, and don't have to be deployed if
that functionality is not used.