PdfGraphics2D.java: Add missing public keyword on constructor
When calling the main PdfGraphics2D constructor, an error was given
saying that it was private. Adding the public keyword fixed this.

Adding a three item constructor, because it is common to just use
the Content, width, and height without other parameters. This
keeps code tidy by not having null, false, false, 0 all the time.

Signed-off by: Jonathan Scruggs (j.scruggs@gmail.com)