fix NPE in PdfWriter#getNewObjectNumber
When trying to add an indirect reference with `PdfWriter`, an NPE occurs because `currentPdfReaderInstance` is null.
This fix ensures that it is using the correct `PdfReaderInstance` via the `importedPages` field.

I'm not entirely sure why `currentPdfReaderInstance` exists at all, since it is immediately reset to `null`
after any methods that use it, but I used it in this fix for consistency's sake.