Check if the annotation belongs to a valid page
The AcroFields class will retrieve "invisible signatures" (and possibly other
types of annotations), which are not present in the page annotation array,
existing only as single level annotations. These are returned with page == -1,
which would cause a NPE at [2] if the fields are to be flattened on close
(i.e.: if stamper.setFormFlattening(true); was called).

[1] https://github.com/LibrePDF/OpenPDF/blob/master/openpdf/src/main/java/com/lowagie/text/pdf/AcroFields.java\#L239
[2] https://github.com/LibrePDF/OpenPDF/blob/master/openpdf/src/main/java/com/lowagie/text/pdf/PdfStamperImp.java\#L880

Signed-off-by: Bruno Melo <bsilva.melo@gmail.com>