Improved owner password recognition
A number of the example AES256R6 encrypted PDFs have a peculiarity: The U and O values by specification are 48 byte values, but in these PDFs they are supplemented with zeroes to 127 bytes each. In most contexts this is no problem as one usually takes n bytes starting at offset m for some calculation. When validating the owner password, though, the whole U value is used. To make this code work for the PDFs in question, one has to use only the leading 48 bytes of the U value.

PdfReader.readDecryptedDocObj has been adapted to do so.
Additionally the DecryptAES256R6Test tests have been adapted to take this into account.