Bugfix - PdfEncryption copy constructor did not copy mkey
Lonzak had reported a NullPointerException in setHashKey during the update of the MD5 hash by mkey when stamping legacy encrypted PDFs.

When stamping encrypted PDFs in append mode, a copy of the PdfEncryption instance of the PdfReader is stored in the PdfStamperImp instance.
The mkey value is used for the legacy encryption algorithms (and only there).
Thus, stamping tests with AES256R6 succeeded while stamping tests with legacy algorithms don't exist...